Interpreter: Support recursive calls
It turns out python scoping is a bit weird. Top level functions are
added to the `globals()` table. Function calls to self or a parent
function which is not on the top level are passed as local variables.
Here is an example:

```python
print("s00 " + str(globals()))
print("s00 " + str(locals()))

s00 = 0

def scope_01(con):
    s01 = 1
    print("s01" + str(globals()))
    print("s01 " + str(locals()))

    def scope_02(con):
        s02 = 2
        print("s02" + str(globals()))
        print("s02 " + str(locals()))

        def scope_03():
            s03 = 3
            print("s03" + str(globals()))
            print("s03 " + str(locals()))
            scope_01(False)
            scope_02(False)

        if con:
            scope_03()
    if con:
        scope_02(True)

scope_01(True)
```

This commit adds a hack that simulates the global namespace. This at
least allows simple recursion.