Fix bug introduced in #39
The sbrc is used to approximate the reference count on a region due to subregions having non-zero lrc.
When I removed the ability to have multiple parent pointers, that broke the invariant, and lead to a leak that Asan detected.

This correctly handles the sbrc when a region is reparented, and
ensures a region is collected when its final reference is a parent pointer.